refactor: centralize format-bytes helper and improve file listing in Edge Storage
